Refactor: 삭제 관련 동작을 internal method로 빼기
